+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.4.0] - 2022-10-07
+
 - Add missing `histogram` and ` distribution` metric types.
 
 ## [0.3.0] - 2021-08-08